refactor(backend): change stats to decimals

+++ apps/backend/src/models.rs
@@ -844,10 +844,10 @@ pub mod fitness {
     pub struct TotalMeasurement {
         /// The number of personal bests achieved.
         pub personal_bests_achieved: usize,
-        pub weight: usize,
-        pub reps: usize,
-        pub distance: usize,
-        pub duration: usize,
+        pub weight: Decimal,
+        pub reps: Decimal,
+        pub distance: Decimal,
+        pub duration: Decimal,
     }
 
     #[derive(
